+++ lib/ooxml_parser/common_parser/common_data/table/table_properties.rb
@@ -68,7 +68,7 @@ module OoxmlParser
         when 'tblpPr'
           table_properties.table_positon = TablePosition.parse(table_props_node_child)
         when 'tblCellMar'
-          table_properties.table_cell_margin = TableMargins.parse(table_props_node_child)
+          table_properties.table_cell_margin = TableMargins.new(parent: table_properties).parse(table_props_node_child)
         end
       end
       table_properties.table_look = TableLook.parse(table_properties_node) if table_properties.table_look.nil?