+++ spec/document/elements/table/cell/cell_properties_spec.rb
@@ -14,7 +14,11 @@ describe OoxmlParser::CellProperties do
   it 'table_custom_cell_margin.docx' do
     docx = OoxmlParser::DocxParser.parse_docx('spec/document/elements/table/cell/properties/table_custom_cell_margin.docx')
     margins_in_doc = docx.element_by_description(location: :canvas, type: :paragraph)[1].rows[0].cells[0].properties.table_cell_margin
-    expect(OoxmlParser::TableMargins.new(false, 0.5, 3.19, 6.0, 3.19)).to eq(margins_in_doc)
+    expect(OoxmlParser::TableMargins.new(false,
+                                         OoxmlParser::OoxmlSize.new(0.5, :centimeter),
+                                         OoxmlParser::OoxmlSize.new(3.19, :centimeter),
+                                         OoxmlParser::OoxmlSize.new(6.0, :centimeter),
+                                         OoxmlParser::OoxmlSize.new(3.19, :centimeter))).to eq(margins_in_doc)
   end
 
   it 'table_cell_shade.docx' do